manipulate.py is now running as long as daataProcessing.py. Enables to terminate process as soon as possible
dataProcessing.py is creating a "Running.txt" file on start and deletes
it after the simulation is finished. manipulateData is checking if the
file "Running.txt" exsists. If so the Script waits X Seconds until it
checks it again and terminates as soon as it is no longer finding the
"Running.txt" file.