Added LogFile to manipulateData.py
Issue: In a later stage this script will run on a server and we won't
start the script by the command line. We might not be able to read the
prints of the script in a proper way. I tried to do it by using POPEN
but it  wasn't succesfull. Writting now all important messages into
dataProcessing.log.