IMPLEMENTED: Exception handling for DataBase Connection
If the database path was set wrong, the script terminated with 0 instead
of 1. To avoid that, an exception handling was implemented. If the
connection is successfull, the log entry "DataBase Connection: Success"
is added and on failure the entry "DataBase Connection: Fail" and the
database path are added to the logfile dataProcessing.log. The script
now also terminates with 1 and returns the message "DataBase Connection
Failed" on connection failure.