IMPLEMENTED: Possibilty for averageEnergyConsumptionView to request data for all iterations
As part of US86, all iterations should also be requestable by the
averageEnergyConsumptionView. Therefor type is implemented as parameter
for to the API. If last10 are requested, the last 10 iterations are
returned, if all are requested all iterations are requested and if
neither of both are requested, the message 'Wrong Type Selected' will be
returned.